QuoteRepository 데이터 레이어

See merge request s13-webmobile4-sub1/S13P11D202!15

+++ GLIM-AOS/app/src/main/java/com/ssafy/glim/core/domain/model/Quote.kt
@@ -5,5 +5,5 @@ data class Quote(
     val text: String,
     val bookTitle: String,
     val page: String? = null,
-    val likes: Int
+    val likes: Long
 )
